fix: resolve return calcularEstatisticasTransacoes

+++ src/main/java/com/joaomauricio/desafio_itau/controller/dtos/EstatisticasResponseDTO.java
@@ -1,7 +1,7 @@
 package com.joaomauricio.desafio_itau.controller.dtos;
 
 public record EstatisticasResponseDTO(
-        Integer count,
+        Long count,
         Double sum,
         Double avg,
         Double min,
